refactor(core): setup single-place layout selection
[no changelog]
