+++ core/site_scons/site_tools/micropython/__init__.py
@@ -39,6 +39,8 @@ def generate(env):
         is_t2b1 = env["TREZOR_MODEL"] == "R"
         backlight = env["backlight"]
         optiga = env["optiga"]
+        layout_tt = env["ui_layout"] == "UI_LAYOUT_TT"
+        layout_tr = env["ui_layout"] == "UI_LAYOUT_TR"
         interim = f"{target[:-4]}.i"  # replace .mpy with .i
         sed_scripts = " ".join(
             [
@@ -46,6 +48,8 @@ def generate(env):
                 rf"-e 's/utils\.BITCOIN_ONLY/{btc_only}/g'",
                 rf"-e 's/utils\.USE_BACKLIGHT/{backlight}/g'",
                 rf"-e 's/utils\.USE_OPTIGA/{optiga}/g'",
+                rf"-e 's/utils\.UI_LAYOUT == \"TT\"/{layout_tt}/g'",
+                rf"-e 's/utils\.UI_LAYOUT == \"TR\"/{layout_tr}/g'",
                 r"-e 's/if TYPE_CHECKING/if False/'",
                 r"-e 's/import typing/# \0/'",
                 r"-e '/from typing import (/,/^\s*)/ {s/^/# /}'",
