+++ core/src/trezor/ui/layouts/homescreen.py
@@ -1,6 +1,6 @@
 from trezor import utils
 
-if utils.INTERNAL_MODEL in ("T2T1", "D001"):
-    from .tt_v2.homescreen import *  # noqa: F401,F403
-elif utils.INTERNAL_MODEL in ("T2B1",):
+if utils.UI_LAYOUT == "TT":
+    from .tt.homescreen import *  # noqa: F401,F403
+elif utils.UI_LAYOUT == "TR":
     from .tr.homescreen import *  # noqa: F401,F403
