+++ core/src/trezor/ui/layouts/recovery.py
@@ -1,6 +1,6 @@
 from trezor import utils
 
-if utils.INTERNAL_MODEL in ("T2T1", "D001"):
-    from .tt_v2.recovery import *  # noqa: F401,F403
-elif utils.INTERNAL_MODEL in ("T2B1",):
+if utils.UI_LAYOUT == "TT":
+    from .tt.recovery import *  # noqa: F401,F403
+elif utils.UI_LAYOUT == "TR":
     from .tr.recovery import *  # noqa: F401,F403
