+++ core/src/trezor/ui/layouts/reset.py
@@ -1,6 +1,6 @@
 from trezor import utils
 
-if utils.INTERNAL_MODEL in ("T2T1", "D001"):
-    from .tt_v2.reset import *  # noqa: F401,F403
-elif utils.INTERNAL_MODEL in ("T2B1",):
+if utils.UI_LAYOUT == "TT":
+    from .tt.reset import *  # noqa: F401,F403
+elif utils.UI_LAYOUT == "TR":
     from .tr.reset import *  # noqa: F401,F403
