+++ core/src/trezor/ui/layouts/tt/fido.py
