+++ core/src/trezor/ui/layouts/tt/homescreen.py
