+++ core/src/trezor/ui/layouts/tt/progress.py
