+++ core/src/trezor/ui/layouts/tt/recovery.py
