+++ core/src/trezor/ui/layouts/tt/reset.py
